alexa_voice_ai/requirements.txt
Flask>=3.0.0
ask-sdk-core==1.19.0
ask-sdk-webservice-support==1.3.3
requests==2.31.0
python-dotenv==1.0.0
anthropic>=0.7.0
